src/requirements.txt
datasets==2.14.4
gensim==4.3.2
html2markdown==0.1.7
ipython==8.12.2
lightning==2.0.8
nltk==3.8.1
numpy==1.24.4
optuna==3.4.0
pandas==2.0.3
plotly==5.17.0
python_Levenshtein==0.23.0
redlines==0.4.2
scikit_learn==1.3.1
scipy==1.11.3
sentence_transformers==2.2.2
torch==2.0.1+cu118
torchmetrics==1.1.1
tqdm==4.66.1
transformers==4.32.1